Rename jballerina initializer class
