+++ tests/ballerina-test-utils/src/main/java/org/ballerinalang/test/listener/JBallerinaTestInitializer.java
@@ -6,17 +6,75 @@
  * in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
+ * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
+
 package org.ballerinalang.test.listener;
 
-public class JBallerinaTestInitializer {
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.testng.ITestContext;
+import org.testng.ITestListener;
+import org.testng.ITestResult;
+
+import static org.ballerinalang.test.util.TestConstant.ENABLE_JBALLERINA_TESTS;
+
+/**
+ * A test suit listener for jballerina test cases initialization.
+ *
+ * @since 0.995.0
+ */
+public class JBallerinaTestInitializer implements ITestListener {
+
+    private static Logger log = LoggerFactory.getLogger(JBallerinaTestInitializer.class);
+
+    @Override
+    public void onStart(ITestContext context) {
+        String property = context.getCurrentXmlTest().getParameter(ENABLE_JBALLERINA_TESTS);
+        if (property != null && Boolean.valueOf(property)) {
+            log.info("JBallerina tests initialized...");
+            System.setProperty(ENABLE_JBALLERINA_TESTS, "true");
+        }
+    }
+
+    @Override
+    public void onFinish(ITestContext context) {
+        String property = context.getCurrentXmlTest().getParameter(ENABLE_JBALLERINA_TESTS);
+        if (property != null && Boolean.valueOf(property)) {
+            log.info("JBallerina tests disabled...");
+            System.clearProperty(ENABLE_JBALLERINA_TESTS);
+        }
+    }
+
+    @Override
+    public void onTestStart(ITestResult result) {
+        //ignore
+    }
+
+    @Override
+    public void onTestSuccess(ITestResult result) {
+        //ignore
+    }
+
+    @Override
+    public void onTestFailure(ITestResult result) {
+        //ignore
+    }
+
+    @Override
+    public void onTestSkipped(ITestResult result) {
+        //ignore
+    }
+
+    @Override
+    public void onTestFailedButWithinSuccessPercentage(ITestResult result) {
+        //ignore
+    }
 }
